Now data is always loaded.
